pyproject.toml
[project]
name = "ai-factory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "pyyaml>=6.0.3",
]
